+++ packages/common/src/projects/defaults.ts
@@ -22,7 +22,6 @@ export const DEFAULT_PROJECT: Partial<IHubProject> = {
     metricDisplays: [],
   },
   features: InitiativeDefaultFeatures,
-  catalogs: [],
 };
 
 /**
@@ -44,7 +43,9 @@ export const DEFAULT_PROJECT_MODEL: IModel = {
   data: {
     display: "about",
     permissions: [],
-    catalogs: [],
+    catalog: {
+      schemaVersion: 0,
+    },
     status: HubEntityStatus.notStarted,
     view: {
       contacts: [],