+++ packages/common/src/sites/defaults.ts
@@ -13,7 +13,6 @@ export const DEFAULT_SITE: Partial<IHubSite> = {
   typeKeywords: ["Hub Site", "hubSite"],
   catalog: { schemaVersion: 0 },
   permissions: [],
-  catalogs: [],
   schemaVersion: 1,
   features: SiteDefaultFeatures,
 };
@@ -35,7 +34,7 @@ export const DEFAULT_SITE_MODEL: IModel = {
     },
   },
   data: {
-    catalogs: [],
+    catalog: { schemaVersion: 0 },
     layout: {},
   },
 } as unknown as IModel;