+++ packages/common/test/content/hostedServiceUtils.test.ts
@@ -4,35 +4,11 @@ import { IHubEditableContent } from "../../src";
 import {
   hasServiceCapability,
   isAGOFeatureServiceUrl,
-  isHostedFeatureServiceEntity,
-  isHostedFeatureServiceItem,
   isHostedFeatureServiceMainEntity,
   isHostedFeatureServiceMainItem,
   ServiceCapabilities,
   toggleServiceCapability,
 } from "../../src/content/hostedServiceUtils";
-
-// NOTE: isHostedFeatureServiceItem is deprecated. Remove this test when the function is removed.
-describe("isHostedFeatureServiceItem", () => {
-  beforeAll(() => {
-    // suppress deprecation warnings
-    // tslint:disable-next-line: no-empty
-    spyOn(console, "warn").and.callFake(() => {});
-  });
-  it("returns true for hosted feature service items", () => {
-    const item = {
-      type: "Feature Service",
-      typeKeywords: ["Hosted Service"],
-    } as IItem;
-
-    expect(isHostedFeatureServiceItem(item)).toBeTruthy();
-  });
-
-  it("returns false for other items", () => {
-    const item = { type: "PDF" } as IItem;
-    expect(isHostedFeatureServiceItem(item)).toBeFalsy();
-  });
-});
 
 describe("isHostedFeatureServiceMainItem", () => {
   it("returns true for hosted feature service main items", () => {
@@ -55,28 +31,6 @@ describe("isHostedFeatureServiceMainItem", () => {
   it("returns false for other items", () => {
     const item = { type: "PDF" } as IItem;
     expect(isHostedFeatureServiceMainItem(item)).toBeFalsy();
-  });
-});
-
-// NOTE: isHostedFeatureServiceEntity is deprecated. Remove this test when the function is removed.
-describe("isHostedFeatureServiceEntity", () => {
-  beforeAll(() => {
-    // suppress deprecation warnings
-    // tslint:disable-next-line: no-empty
-    spyOn(console, "warn").and.callFake(() => {});
-  });
-  it("returns true for hosted feature service content entities", () => {
-    const entity = {
-      type: "Feature Service",
-      typeKeywords: ["Hosted Service"],
-    } as IHubEditableContent;
-
-    expect(isHostedFeatureServiceEntity(entity)).toBeTruthy();
-  });
-
-  it("returns false for other content entities", () => {
-    const entity = { type: "PDF" } as IHubEditableContent;
-    expect(isHostedFeatureServiceEntity(entity)).toBeFalsy();
   });
 });