+++ packages/common/test/pages/_internal/PageUiSchemaEdit.test.ts
@@ -68,6 +68,12 @@ describe("buildUiSchema: page edit", () => {
                     icon: true,
                     labelKey: `some.scope.fields.name.maxLengthError`,
                   },
+                  {
+                    type: "ERROR",
+                    keyword: "format",
+                    icon: true,
+                    labelKey: `some.scope.fields.name.entityTitleValidatorError`,
+                  },
                 ],
               },
             },