+++ packages/common/test/sites/HubSite.test.ts
@@ -14,7 +14,7 @@ import {
   IVersionMetadata,
   getProp,
 } from "../../src";
-import { Catalog } from "../../src/search";
+import { Catalog } from "../../src/search/Catalog";
 import * as ContainsModule from "../../src/core/_internal/deepContains";
 import * as EnrichEntityModule from "../../src/core/enrichEntity";
 
@@ -300,7 +300,7 @@ describe("HubSite Class:", () => {
       );
       // pass in a project catalog
       const result = await chk.contains("cc0", [
-        { id: "4ef", entityType: "item", catalog: createCatalog("00b") },
+        { id: "4ef", hubEntityType: "project", catalog: createCatalog("00b") },
       ]);
       expect(containsSpy).toHaveBeenCalledTimes(1);
       const hiearchy = containsSpy.calls.argsFor(0)[2];
@@ -345,9 +345,9 @@ describe("HubSite Class:", () => {
         authdCtxMgr.context
       );
       // First call will warm the cache
-      await chk.contains("cc0", [{ id: "4ef", entityType: "item" }]);
+      await chk.contains("cc0", [{ id: "4ef", hubEntityType: "project" }]);
       // second call will use the cache
-      await chk.contains("cc1", [{ id: "4ef", entityType: "item" }]);
+      await chk.contains("cc1", [{ id: "4ef", hubEntityType: "project" }]);
       expect(containsSpy).toHaveBeenCalledTimes(2);
       // verify first call does not send the 4ef catalog
       const hiearchy = containsSpy.calls.argsFor(0)[2];
@@ -378,7 +378,6 @@ describe("HubSite Class:", () => {
         catalog: {
           schemaVersion: 0,
         },
-        catalogs: [],
         permissions: [],
         settings: {
           features: {
@@ -478,7 +477,6 @@ describe("HubSite Class:", () => {
         created: 123456,
         creator: "paige_pa",
         data: {
-          catalogs: [],
           settings: {
             features: {
               "hub:site:content": true,